requirements.txt
anyio==3.6.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
biopython
black==22.6.0
bleach==5.0.1
boto3==1.24.33
botocore
Cerberus==1.3.2
certifi==2021.10.8
cffi==1.15.1
chardet==4.0.0
click==8.1.3
cmapPy==4.0.1
cryptography==37.0.4
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
docutils==0.19
elucidatacloudpathlib==0.6.6
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.1
fastobo==0.12.1
h5py==3.7.0
httplib2==0.20.4
idna==2.10
iniconfig==1.1.1
interchange==2021.0.4
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
json5==0.9.8
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.6
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.30
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
neo4j==4.4.5
nest-asyncio==1.5.5
networkx==2.8.5
notebook-shim==0.1.0
numpy==1.23.1
packaging==21.3
pandas
pandocfilters==1.5.0
pansi==2020.7.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
pluggy==1.0.0
polly-python
postpy2==0.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.30
pronto==2.5.0
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py2neo==2021.2.3
pycparser==2.21
Pygments==2.12.0
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
python-dateutil
python-dotenv==0.20.0
python-magic==0.4.24
pytz==2021.1
PyYAML==6.0
pyzmq==23.2.0
requests==2.25.1
retrying==1.3.3
rst2txt==1.1.0
s3transfer==0.6.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
sqlparse==0.4.2
stack-data==0.3.0
terminado==0.15.0
tinycss2==1.1.1
tomli==2.0.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
wrapt==1.14.1
